refactor: extract funding round utility functions into separate module
