+++ src/app/funding-rounds/[id]/page.tsx
@@ -23,6 +23,12 @@ import { Button } from '@/components/ui/button'
 import { FundingRoundService } from '@/services'
 import prisma from '@/lib/prisma'
 import { notFound } from 'next/navigation'
+import {
+	formatDate,
+	getPreviousAndNextForBetweenPhase,
+	getTimeRemaining,
+	getTimeSince,
+} from '@/lib/funding-round-utils'
 
 export const dynamic = 'force-dynamic'
 
@@ -35,13 +41,6 @@ const FUNDING_ROUND_PHASES: Exclude<FundingRoundPhase, 'UPCOMING'>[] = [
 	'BETWEEN_PHASES',
 ]
 
-type FundingRoundDashboardProps = {
-	params: Promise<{ id: string }>
-}
-
-const formatDate = (date: Date) =>
-	`${date.toLocaleString('default', { month: 'short' })} ${date.getDate()}, ${date.getFullYear()}`
-
 const getFundingRoundById = async (
 	id: string,
 ): Promise<FundingRoundWithPhases | null> => {
@@ -51,7 +50,9 @@ const getFundingRoundById = async (
 
 export default async function FundingRoundDashboard({
 	params,
-}: FundingRoundDashboardProps) {
+}: {
+	params: Promise<{ id: string }>
+}) {
 	const { id } = await params
 
 	const data = await getFundingRoundById(id)
@@ -302,90 +303,5 @@ function FundingRoundPhaseComponent({
 			return <CompletedPhase />
 		default:
 			return null
-	}
-}
-
-function formatDateDiff(diff: number): string {
-	// Calculate days, hours, minutes
-	const days = Math.floor(diff / (1000 * 60 * 60 * 24))
-	const hours = Math.floor((diff % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60))
-	const minutes = Math.floor((diff % (1000 * 60 * 60)) / (1000 * 60))
-
-	if (days > 0) {
-		return `${days}d ${hours}h`
-	}
-	if (hours > 0) {
-		return `${hours}h ${minutes}m`
-	}
-	return `${minutes}m`
-}
-
-function getTimeRemaining(date: Date): string {
-	const now = Date.now()
-	const diff = date.getTime() - now
-
-	// Time already passed
-	if (diff < 0) {
-		return 'Ended'
-	}
-
-	return formatDateDiff(diff)
-}
-
-function getTimeSince(date: Date): string {
-	const now = Date.now()
-	const diff = now - date.getTime()
-
-	return formatDateDiff(diff)
-}
-
-const getPreviousAndNextForBetweenPhase = (
-	phases: FundingRoundPhases,
-): {
-	nextPhase: {
-		name: FundingRoundPhase
-		startDate: string
-		endDate: string
-	} | null
-	previousPhase: {
-		name: FundingRoundPhase
-		startDate: string
-		endDate: string
-	} | null
-} => {
-	const now = new Date()
-
-	const destructuredPhases = Object.entries(phases).map(
-		([name, { startDate, endDate }], index) => ({
-			index,
-			name,
-			startDate,
-			endDate,
-		}),
-	)
-
-	const nextPhase = destructuredPhases.find(
-		({ startDate }) => new Date(startDate) > now,
-	)
-
-	const previousPhase = destructuredPhases.find(
-		({ endDate }) => new Date(endDate) < now,
-	)
-
-	return {
-		nextPhase: nextPhase
-			? {
-					name: nextPhase.name as FundingRoundPhase,
-					startDate: nextPhase.startDate,
-					endDate: nextPhase.endDate,
-				}
-			: null,
-		previousPhase: previousPhase
-			? {
-					name: previousPhase.name as FundingRoundPhase,
-					startDate: previousPhase.startDate,
-					endDate: previousPhase.endDate,
-				}
-			: null,
 	}
 }
